Move typescript defs out of tsx file (no jsx)
