--- apps/site/assets/ts/projects/components/Image.tsx
@@ -1,5 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export interface Image {
-  url: string;
-  alt: string;
-}
