--- apps/site/assets/ts/projects/components/Route.tsx
@@ -1,6 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export interface Route {
-  mode: string;
-  id: string;
-  group: string;
-}
